third_party/WebKit/LayoutTests/fast/forms/old-names.html

<!DOCTYPE html>
<html>
<head>
<meta http-equiv="X-UA-Compatible" content="IE=edge" />
<script src="../js/resources/js-test-pre.js"></script>
<script>
function runTest()
{
    description("This tests accessing form elements by name. "
        + "IE updates collections immediatelly after name changes, but A "
        + "collection never be changed to a single node when the number of "
        + "matched nodes becomes 1. "
        + "Firefox remembers every name item has been accessed with, but "
        + "forgets items that have not been accessed. "
        + "We follow the standard behavior. i.e. collections are updated "
        + "immediately, and it's changed to a single node when the size becomes 1.");

    form = document.getElementById('form');
    a = document.getElementById('a');
    b = document.getElementById('b');

    shouldBe('form.length', '2');
    shouldBe('form.original', 'a');
    shouldBe('form.originalB', 'b');
    shouldBe('form.second', 'undefined');
    shouldBe('form.third', 'undefined');
    shouldBe('form.elements.original', 'a');
    shouldBe('form.elements.originalB', 'b');
    shouldBe('form.elements.second', 'undefined');
    shouldBe('form.elements.third', 'undefined');

    debug('');   
    debug("change the form item a's name to thisWillBeRemembered");
    debug('');   
    a.name="thisWillBeRemembered";
    debug("get the variable value through form element");
    shouldBe('form.thisWillBeRemembered', 'a');
    debug('');   
    debug("now change the form item a's name to thisWillBeRememberedToo");
    debug("access it in boolean context");
    a.name="thisWillBeRememberedToo";
    debug('');   
    if (form.thisWillBeRememberedToo)
        debug('accessed form.thisWillBeRememberedToo');
    debug('');   
    debug("now change the form item a's name to thisWillBeForgotten");
    debug('');   
    a.name="thisWillBeForgotten";
    debug("get the variable value through collection");
    shouldBe('form.elements.thisWillBeForgotten', 'a');
    debug('');   
    debug("now change the form item a's name to thisWillBeForgottenToo, but don't access it afterwards");
    a.name="thisWillBeForgottenToo";
        
    debug('');    
    debug("now change the form item a's name to second");
    debug('');
    a.name="second";
    
    shouldBe('form.length', '2');
    shouldBeUndefined('form.original');
    shouldBe('form.originalB', 'b');
    shouldBe('form.second', 'a');
    shouldBe('form.third', 'undefined');
    shouldBe('form.elements.original', 'undefined');
    shouldBe('form.elements.originalB', 'b');
    shouldBe('form.elements.second', 'a');
    shouldBe('form.elements.third', 'undefined');

    debug('');
    debug("now change the form item a's name to third");
    debug('');

    a.name="third";

    shouldBe('form.length', '2');
    shouldBeUndefined('form.original');
    shouldBe('form.originalB', 'b');
    shouldBeUndefined('form.second');
    shouldBe('form.third', 'a');
    shouldBe('form.elements.original', 'undefined');
    shouldBe('form.elements.originalB', 'b');
    shouldBe('form.elements.second', 'undefined');
    shouldBe('form.elements.third', 'a');

    debug('');
    debug("now change form item b's name to second");
    debug('');

    b.name="second";

    shouldBe('form.length', '2');
    shouldBeUndefined('form.original');
    shouldBeUndefined('form.originalB');
    shouldBe('form.second', 'b');
    shouldBe('form.elements.original', 'undefined');
    shouldBe('form.elements.originalB', 'undefined');
    shouldBe('form.elements.second', 'b');

    debug('');
    debug("now change a form item b's name to third");
    debug('');

    form.second.name = 'third';

    shouldBe('form.length', '2');
    shouldBeUndefined('form.original');
    shouldBeUndefined('form.originalB');
    shouldBeUndefined('form.second');
    shouldBe('form.third.length', '2');
    shouldBe('form.third[0]', 'a');
    shouldBe('form.third[1]', 'b');
    shouldBe('form.elements.original', 'undefined');
    shouldBe('form.elements.originalB', 'undefined');
    shouldBe('form.elements.second', 'undefined');
    shouldBe('form.elements.third.length', '2');
    shouldBe('form.elements.third[0]', 'a');
    shouldBe('form.elements.third[1]', 'b');
    
    debug('');
    debug("now change a form item b's name to fourth");
    debug('');

    form.third[1].name = 'fourth';
    
    shouldBe('form.third', 'a');
    shouldBe('form.third.length', 'undefined');
    shouldBe('form.elements.third', 'a');
    shouldBe('form.elements.third.length', 'undefined');

    debug('');
    debug("now remove element a");
    debug('');

    form.removeChild(a);

    shouldBe('form.length', '1');
    shouldBeUndefined('form.original');
    shouldBeUndefined('form.originalB');
    shouldBeUndefined('form.second');
    shouldBeUndefined('form.third');
    shouldBe('form.fourth', 'b');
    shouldBe('form.elements.original', 'undefined');
    shouldBe('form.elements.originalB', 'undefined');
    shouldBe('form.elements.second', 'undefined');
    shouldBe('form.elements.third', 'undefined');
    shouldBe('form.elements.fourth', 'b');
 
    debug('');
    debug("check we forget names we should");
    debug('');

    shouldBe('form.thisWillBeForgotten', 'undefined');
    shouldBe('form.thisWillBeForgottenToo', 'undefined');
    shouldBeUndefined('form.thisWillBeRemembered');
    shouldBeUndefined('form.thisWillBeRememberedToo');
    debug('');
}
</script>
</head>
<body>
<form id='form'>
<input type='hidden' id='a' name='original'>
<input type='hidden' id='b' name='originalB'>
</form>
<p id="description"></p>
<div id="console"></div>
<script>runTest();</script>
<script src="../js/resources/js-test-post.js"></script>
</body>
</html>
